decoupled polygon and line rasterization to handle GeometryCollection
